fixed bugs with simd implementation, algorithm may be unstable

+++ utilities.cpp
@@ -1,5 +1,6 @@
 #include <arm_neon.h>
 
+#include <iostream>
 #include <memory>
 #include <new>
 #include <random>
@@ -9,7 +10,7 @@
 void init_matricies(float16_t* A, int size) {
     std::random_device rd;
     std::mt19937 gen(rd());
-    std::uniform_real_distribution<float16_t> uniform_dist(-1., 1.);
+    std::uniform_real_distribution<float> uniform_dist(-1., 1.);
     for (int i = 0; i < size; i++)
-        A[i] = uniform_dist(gen);
+        A[i] = (float16_t) uniform_dist(gen);
 }
